+++ Ambient.Saga.Engine.Tests/IntegrationTests/Cqrs/SagaE2EStoryTests.cs
@@ -262,7 +262,7 @@ public class SagaE2EStoryTests : IDisposable
             Id = Guid.NewGuid(),
             ArchetypeRef = "Warrior",
             DisplayName = name,
-            BlockOwnership = new Dictionary<string, int>()
+            BlockOwnership = new Dictionary<string, float>()
         };
 
         AvatarSpawner.SpawnFromModelAvatar(
